+++ provider.go
@@ -1,11 +1,22 @@
 package beans
 
+import "reflect"
+
 func(c *ProviderContainer) AddProvider(fun interface{}) {
-	c.createBean(fun,false)
+	c.createBean(fun,false,false)
 }
 
-func(c *ProviderContainer) createBean(fun interface{},singleton bool) {
-	returnType,beanData,err := newBean(fun,c,singleton)
+func(c *ProviderContainer) createBean(data interface{},singleton,obj bool,cleanUpFunc ...CleanUpFunc) {
+	var (
+		returnType reflect.Type
+		beanData *bean
+		err error
+	)
+	if obj {
+		returnType,beanData,err = newBeanFromObjectData(data,cleanUpFunc...)
+	}else {
+		returnType,beanData,err = newBean(data,c,singleton)
+	}
 	if err != nil {
 		panic(err.Error())
 	}
@@ -17,7 +28,17 @@ func(c *ProviderContainer) createBean(fun interface{},singleton bool) {
 }
 
 func(c *ProviderContainer) AddProviderSingleton(fun interface{}) {
-	c.createBean(fun,true)
+	c.createBean(fun,true,false)
+}
+
+func(c *ProviderContainer) AddObjectSingleton(obj interface{},cleanUpFunc ...CleanUpFunc) {
+	c.createBean(obj,true,true,cleanUpFunc...)
+}
+
+func(c *ProviderContainer) CleanUp() {
+	for _,v := range c.dependencyInitiator {
+		v.cleanUp()
+	}
 }